refactor: Overhaul palette state management and fix critical bugs
Refactored the palette editing feature to use a more robust and scalable architecture with isolated widgets and correct state handling.

- Decoupled UI and data logic by splitting PaletteDetailPage into smaller, focused widgets.
- Corrected a fundamental flaw in the data loading logic where Hive keys were not being assigned to palette objects, which was the root cause of the unresponsive edit/delete buttons.
- Fixed multiple race conditions and state synchronization issues that were preventing changes from being saved reliably.
- Implemented a deep copy in the palette model to ensure true immutability.
- The palette editing and deleting functionality is now working correctly.

+++ lib/project_overview_page.dart
@@ -69,7 +69,7 @@ class HomePage extends ConsumerWidget {
     );
 
     if (confirmed == true) {
-      await ref.read(projectListProvider.notifier).deleteProject(key);
+      await ref.read(projectListProvider.notifier).deleteProject(key as int);
     }
   }
 
@@ -148,7 +148,7 @@ class HomePage extends ConsumerWidget {
                       color: Colors.white,
                       tooltip: 'Delete Project',
                       onPressed: () =>
-                          _deleteProject(context, ref, project.key),
+                          _deleteProject(context, ref, project.key as int),
                     ),
                   ),
                   child: Image.memory(project.thumbnailData, fit: BoxFit.cover),
